+++ package.json
@@ -7,7 +7,7 @@
     "lint": "eslint . --fix",
     "test": "yarn start-sandbox && yarn test-contracts ; yarn kill-sandbox",
     "test-contracts": "jest --runInBand",
-    "generate-api": "ts-node -r tsconfig-paths/register ./packages/minter-cli/src/index.ts --g ./contracts/bin ./contracts/types",
+    "generate-types": "ts-node -r tsconfig-paths/register ./packages/minter-cli/src/index.ts --g ./contracts/bin ./contracts/types",
     "start-sandbox": "flextesa/start-sandbox.sh",
     "kill-sandbox": "flextesa/kill-sandbox.sh",
     "build:watch": "tsc -w -p .",
