+++ packages/contract-api-generator/src/example/example-contract-type-1.ts
