+++ packages/contract-api-generator/src/example/example-contract-type-2.ts
