+++ packages/contract-api-generator/src/generator/common.ts
@@ -1,0 +1,6 @@
+export class GenerateApiError implements Error {
+    public name = `GenerateApiError`;
+    constructor(public message: string, public readonly data: unknown) {
+        console.error(`❌ GenerateApiError: ${message}`, data);
+    }
+}
